Don't add signal handlers until socket connection
If we add the OS signal handlers (via `handler->start()`) before the
client connection to the `QTcpServer` has been completed, we will
deadlock by writing to a socket that doesn't exist yet.

So instead, we emit a Qt Signal once the client has been connected to
the server in socketpair.cpp. unixsignals.cpp then connects that signal
to its `start()` function as a slot. Then no matter what, the signal
handler will not be installed until we are able to write to the socket.
If we receive an OS signal before that, the default OS signal handler
for the signal will be used instead. For example, if we receive SIGTERM
before that, then the process will exit.