src/socketpair.*: log all connection errors
QTcpSocket and QTcpServer emit an Qt signal on error, so we create slots
to receive those signals, and print the error that ocurred to stderr
through qCritical. qCritical gives the admin flexibility in choosing
whether or not to make those errors fatal or not through setting or not
the environment variable `QT_FATAL_CRITICALS`. See
https://doc.qt.io/qt-5/qtglobal.html#qCritical